+++ src/providers/openlibrary/index.ts
@@ -1,3 +1,4 @@
+export * from "./author.js";
 export * from "./book.js";
 export * from "./client.js";
 export * from "./constants.js";